backend/requirements.txt
# Core dependencies
Django==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# CORS headers for cross-origin requests
django-cors-headers==4.3.1

# Environment variables
python-dotenv==1.0.1

# Development and deployment
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest-django==4.7.0

# Utilities
Pillow==10.2.0  # For image processing
requests==2.31.0
